moved page class to pages package
