--- src/test/java/pages/US9_.java
@@ -1,8 +1,0 @@
-package pages;
-
-public class US9_{
-
-//
-// I M HERE
-
-}
